Huge performance boost with searching youtube
Switched the package I was using for youtube search. This new package allows me to limit the number of search items, greatly boosting the performance. The old package lagged the bot so badly that it would cut out every time something was searched. This new package barely lags the bot.